Add support for using application gateway for ingress
This adds support for using an application gateway for ingressing traffic into the cluster. The Application Gateway Ingress Controller, which will be deployed into the cluster, is granted the `Contributor` role to modify the gateway as it sees fit.

Note that the controller will need to be granted the `Microsoft.Network/publicIPAddresses/read` permission on the IP address associated with the gateway for any `ingress` Kubernetes resources to have their address correctly updated.

This also adds support for setting the O/S disk type to `Ephemeral`.